+++ src/filesystem.rs
@@ -7,6 +7,9 @@ use std::io;
 use std::os::unix::fs::{FileTypeExt, PermissionsExt};
 use std::path::{Path, PathBuf};
 
+#[cfg(windows)]
+use faccess::PathExt as _;
+
 use normpath::PathExt;
 
 use crate::dir_entry;
@@ -42,13 +45,13 @@ pub fn is_existing_directory(path: &Path) -> bool {
 }
 
 #[cfg(any(unix, target_os = "redox"))]
-pub fn is_executable(md: &fs::Metadata) -> bool {
+pub fn is_executable(_: &Path, md: &fs::Metadata) -> bool {
     md.permissions().mode() & 0o111 != 0
 }
 
 #[cfg(windows)]
-pub fn is_executable(_: &fs::Metadata) -> bool {
-    false
+pub fn is_executable(path: &Path, _: &fs::Metadata) -> bool {
+    path.executable()
 }
 
 pub fn is_empty(entry: &dir_entry::DirEntry) -> bool {